requirements.txt
websocket-client
kafka-python
psycopg2-binary
